@@ -1,35 +1,73 @@
-import React, { Component } from 'react';
+import React, {useState, useRef, useCallback, useEffect, useMemo, useLayoutEffect} from "react";
 import { Table } from '../Table';
 import './styles.css';
+import TableBody from "../Table/TableBody";
 
-class DataGridBody extends Component {
+const DataGridBody= (props)=> {
+  const {rows, ...props0 } = props;
+  const {height}=props;
 
-  prepareRowsData(rows) {
-    return rows.map(row => this.props.fields.map(field => row[field]));
+const  prepareRowsData=(rows)=> {
+    return rows.map(row => props.fields.map(field => row[field]));
+  }
+   // const rows0=[...prepareRowsData(rows)];
+   // console.log("[rops",props0);
+   // const [rowsRaw,setRowsRaw]=useState([...prepareRowsData(rows)]);
+
+const get_RowsRaw=()=>{
+    const rows0=[...prepareRowsData(rows)];
+    return rows0;
+}
+
+/*
+    useEffect(()=>{
+
+        console.log("gridbody",rowsRaw);
+       // setRowsRaw(rows0);
+    },[rowsRaw])
+*/
+    useEffect(()=>{
+        scrolHeader();
+
+    })
+
+    function scrolHeader(){
+        const target = document.querySelector(".data-grid__body");
+        const elemsToSync = document.querySelector(".data-grid__header");
+
+        target.addEventListener('scroll', (event) => {
+
+//console.log(target.scrollLeft);
+
+            elemsToSync.style.marginLeft = -target.scrollLeft + 'px';
+
+
+        });
+
+    }
+
+const  handleOnRowClick=(index)=> {
+    console.log(rows[index]);
   }
 
-  handleOnRowClick(index) {
-    console.log(this.props.rows[index]);
+const  handleOnRowDoubleClick=(index)=> {
+    console.log(rows[index]);
   }
 
-  handleOnRowDoubleClick(index) {
-    console.log(this.props.rows[index]);
-  }
 
-  render() {
-    const { rows, ...props } = this.props;
+
     return (
-      <div className="data-grid__body">
-        <Table
-          data={this.prepareRowsData(rows)}
-          {...props}
+      <div className="data-grid__body " style={{  height: height+"px" }}>
+        <TableBody
+          rows={rows}
+          {...props0}
           singleSelect
-          onRowClick={this.handleOnRowClick.bind(this)}
-          onRowDoubleClick={this.handleOnRowDoubleClick.bind(this)}
+          onRowClick={handleOnRowClick}
+          onRowDoubleClick={handleOnRowDoubleClick}
         />
       </div>
     );
-  }
+
 }
 
 export default DataGridBody;
