@@ -1,17 +1,18 @@
 import React, { Component } from 'react';
 import { Table } from '../Table';
-import './styles.css';
+//import './styles.css';
 
-class DataGridFooter extends Component {
-  render() {
-    const { rows, ...props } = this.props;
+const DataGridFooter =(props)=> {
+  const { rows, ...props0 } = props;
+
+
 
     return (
       <div className="data-grid__footer">
-        <Table data={rows} {...props} />
+        <Table data={rows} {...props0}  />
       </div>
     );
-  }
+
 }
 
 export default DataGridFooter;
